falta mapear

+++ includes/fdf.h
@@ -6,7 +6,7 @@
 /*   By: adpachec <adpachec@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/14 11:37:54 by adpachec          #+#    #+#             */
-/*   Updated: 2023/02/08 11:40:23 by adpachec         ###   ########.fr       */
+/*   Updated: 2023/02/14 11:37:51 by adpachec         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define FDF_H
 
 # include "unistd.h"
-//# include "mlx.h"
+# include "mlx.h"
 # include "math.h"
 # include <unistd.h>
 # include <stdlib.h>
@@ -57,15 +57,15 @@ typedef struct s_map_proj
 
 typedef struct s_coord
 {
-	long int	x1;
-	long int	y1;
-	long int	x2;
-	long int	y2;
-	int			i;
-	int			j;
-	float		x_inc;
-	float		y_inc;
-}				t_coord;
+	float	x1;
+	float	y1;
+	float	x2;
+	float	y2;
+	int		i;
+	int		j;
+	float	x_inc;
+	float	y_inc;
+}			t_coord;
 
 char	*get_next_line(int fd);
 char	*ft_strjoin_gnl(char *line, char **save_read);
